specs for editing and delete articles

+++ app/controllers/articles_controller.rb
@@ -26,22 +26,44 @@ class ArticlesController < ApplicationController
   end
   
   def edit
-  end
-  
-  def update 
-    if @article.update(article_params)
-      flash[:success] = "Article has been update"
-      redirect_to @article
-    else
-      flash.now[:danger] = "Article has not been updated"
-      render :edit
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article."
+      redirect_to root_path
     end
   end
   
-  def destroy
-    if @article.destroy
-      flash[:success]  = "Article has been deleted"
+  def update 
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article."
+      redirect_to root_path
+    else
+      if @article.update(article_params)
+        flash[:success] = "Article has been update"
+        redirect_to @article
+      else
+        flash.now[:danger] = "Article has not been updated"
+        render :edit
+      end
+    end
+  end
+  
+  def destroy 
+    puts "esta en destroy"
+    unless @article.user == current_user
+    puts 1
+      flash[:danger] = "You can only delete your own article."
       redirect_to articles_path
+    else
+      puts 2
+      if @article.destroy
+        puts 3
+        flash[:success] = "Article has been deleted"
+        redirect_to articles_path
+      else
+        puts 4
+        flash.now[:danger] = "Article has not been deleted"
+        redirect_to articles_path
+      end
     end
   end
